Move MyClient & rename to VCRolesClient
