+++ cogs/all-link.py
@@ -4,12 +4,12 @@ import discord
 from discord import app_commands
 from discord.ext import commands
 
-from bot import MyClient
 from utils.checks import check_any, command_available, is_owner
+from utils.client import VCRolesClient
 
 
 class AllLink(commands.Cog):
-    def __init__(self, client: MyClient):
+    def __init__(self, client: VCRolesClient):
         self.client = client
 
     all_commands = app_commands.Group(
@@ -245,5 +245,5 @@ class AllLink(commands.Cog):
         return self.client.incr_counter("all_reverse_unlink")
 
 
-async def setup(client: MyClient):
+async def setup(client: VCRolesClient):
     await client.add_cog(AllLink(client))
