+++ cogs/linking-commands.py
@@ -4,13 +4,13 @@ import discord
 from discord import app_commands
 from discord.ext import commands
 
-from bot import MyClient
 from utils.checks import check_any, command_available, is_owner
+from utils.client import VCRolesClient
 from utils.linking import LinkingUtils
 
 
 class Linking(commands.Cog):
-    def __init__(self, client: MyClient):
+    def __init__(self, client: VCRolesClient):
         self.client = client
         self.linking = LinkingUtils(client)
 
@@ -154,5 +154,5 @@ class Linking(commands.Cog):
         return self.client.incr_counter("reverse_unlink")
 
 
-async def setup(client: MyClient):
+async def setup(client: VCRolesClient):
     await client.add_cog(Linking(client))
