+++ cogs/logging.py
@@ -4,12 +4,12 @@ import discord
 from discord import app_commands
 from discord.ext import commands
 
-from bot import MyClient
 from utils.checks import check_any, command_available, is_owner
+from utils.client import VCRolesClient
 
 
 class Logging(commands.Cog):
-    def __init__(self, client: MyClient):
+    def __init__(self, client: VCRolesClient):
         self.client = client
 
     @app_commands.command()
@@ -59,5 +59,5 @@ class Logging(commands.Cog):
         return self.client.incr_counter("logging")
 
 
-async def setup(client: MyClient):
+async def setup(client: VCRolesClient):
     await client.add_cog(Logging(client))
