+++ cogs/permanent-link.py
@@ -4,13 +4,13 @@ import discord
 from discord import app_commands
 from discord.ext import commands
 
-from bot import MyClient
 from utils.checks import check_any, command_available, is_owner
+from utils.client import VCRolesClient
 from utils.linking import LinkingUtils
 
 
 class PermLink(commands.Cog):
-    def __init__(self, client: MyClient):
+    def __init__(self, client: VCRolesClient):
         self.client = client
         self.linking = LinkingUtils(client)
 
@@ -181,5 +181,5 @@ class PermLink(commands.Cog):
         return self.client.incr_counter("perm_reverse_unlink")
 
 
-async def setup(client: MyClient):
+async def setup(client: VCRolesClient):
     await client.add_cog(PermLink(client))
