+++ cogs/ping.py
@@ -2,11 +2,11 @@ import discord
 from discord import app_commands
 from discord.ext import commands
 
-from bot import MyClient
+from utils.client import VCRolesClient
 
 
 class Ping(commands.Cog):
-    def __init__(self, client: MyClient):
+    def __init__(self, client: VCRolesClient):
         self.client = client
 
     @app_commands.command()
@@ -20,5 +20,5 @@ class Ping(commands.Cog):
         return self.client.incr_counter("ping")
 
 
-async def setup(client: MyClient):
+async def setup(client: VCRolesClient):
     await client.add_cog(Ping(client))
