+++ cogs/vc-control.py
@@ -5,12 +5,12 @@ import discord
 from discord import app_commands
 from discord.ext import commands
 
-from bot import MyClient
 from utils.checks import check_any, command_available, is_owner
+from utils.client import VCRolesClient
 
 
 class VCControl(commands.Cog):
-    def __init__(self, client: MyClient):
+    def __init__(self, client: VCRolesClient):
         self.client = client
 
     control_commands = app_commands.Group(
@@ -172,5 +172,5 @@ class VCControl(commands.Cog):
         return self.client.incr_counter("vc_undeafen")
 
 
-async def setup(client: MyClient):
+async def setup(client: VCRolesClient):
     await client.add_cog(VCControl(client))
