+++ cogs/voice-generators.py
@@ -4,13 +4,13 @@ import discord
 from discord import app_commands
 from discord.ext import commands
 
-from bot import MyClient
 from utils.checks import check_any, command_available, is_owner
+from utils.client import VCRolesClient
 from views.interface import Interface
 
 
 class VoiceGen(commands.Cog):
-    def __init__(self, client: MyClient):
+    def __init__(self, client: VCRolesClient):
         self.client = client
 
     async def remove_generator(self, data: dict[str, Union[str, list]]):
@@ -182,5 +182,5 @@ class VoiceGen(commands.Cog):
         return self.client.incr_counter("voice_generator_remove")
 
 
-async def setup(client: MyClient):
+async def setup(client: VCRolesClient):
     await client.add_cog(VoiceGen(client))
